:recycle: extract helpers from test-case to builders
